requirements.txt
asgiref==3.8.1
boto3==1.35.26
botocore==1.35.26
Brotli==1.1.0
certifi==2024.8.30
cffi==1.17.1
cloudinary==1.41.0
cssselect2==0.7.0
Django==5.1
django-storages==1.14.4
fonttools==4.53.1
html5lib==1.1
jmespath==1.0.1
packaging==24.1
pillow==10.4.0
pycparser==2.22
pydyf==0.11.0
pyphen==0.16.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
s3transfer==0.10.2
six==1.16.0
sqlparse==0.5.1
tinycss2==1.3.0
urllib3==2.2.3
weasyprint==62.3
webencodings==0.5.1
whitenoise==6.7.0
zopfli==0.2.3
